refactor(pam): move the gated-collection indicator token into a pam folder
Groups it with the other PAM-specific files under vault-filter/shared/components
instead of sitting alongside the OSS sidebar components, matching how other
PAM additions from this UAT stack are organized.
