+++ apps/web/src/app/vault/individual-vault/vault-filter/shared/components/vault-filter-section.component.spec.ts
@@ -17,7 +17,7 @@ import {
 
 import { CoachmarkService } from "../../../../components/coachmark";
 
-import { VAULT_FILTER_GATED_COLLECTION_INDICATOR } from "./vault-filter-gated-collection-indicator.token";
+import { VAULT_FILTER_GATED_COLLECTION_INDICATOR } from "./pam/vault-filter-gated-collection-indicator.token";
 import { VaultFilterSectionComponent } from "./vault-filter-section.component";
 
 @Component({
